+++ src/main/java/com/hbm/blocks/ModBlocks.java
@@ -1256,11 +1256,11 @@ public class ModBlocks {
     public static final Block sat_resonator = new DecoBlock(Material.IRON, "deco_sat_resonator").setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.blockTab);
 
     //Rad'nts
-    public static final Block absorber = new BlockBakeBase(Material.IRON, "absorber").setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
-    public static final Block absorber_red = new BlockBakeBase(Material.IRON, "absorber_red").setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
-    public static final Block absorber_green = new BlockBakeBase(Material.IRON, "absorber_green").setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
-    public static final Block absorber_pink = new BlockBakeBase(Material.IRON, "absorber_pink").setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
     public static final Block rad_absorber = new BlockAbsorber(Material.IRON, "rad_absorber").setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
+    @Deprecated public static final Block absorber = new BlockBakeOld(Material.IRON, "absorber", rad_absorber.getStateFromMeta(BlockAbsorber.EnumAbsorberTier.BASE.ordinal())).setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
+    @Deprecated public static final Block absorber_red = new BlockBakeOld(Material.IRON, "absorber_red", rad_absorber.getStateFromMeta(BlockAbsorber.EnumAbsorberTier.RED.ordinal())).setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
+    @Deprecated public static final Block absorber_green = new BlockBakeOld(Material.IRON, "absorber_green", rad_absorber.getStateFromMeta(BlockAbsorber.EnumAbsorberTier.GREEN.ordinal())).setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
+    @Deprecated public static final Block absorber_pink = new BlockBakeOld(Material.IRON, "absorber_pink", rad_absorber.getStateFromMeta(BlockAbsorber.EnumAbsorberTier.PINK.ordinal())).setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
     public static final Block decon = new BlockDeconRad(Material.IRON, "decon", 0.5F).setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);
     public static final Block decon_digamma = new BlockDeconDi(Material.IRON, "decon_digamma", 0.001F).setHardness(5.0F).setResistance(10.0F).setCreativeTab(MainRegistry.machineTab);